requirements.txt
aniso8601==9.0.1
autopep8==1.5.7
click==8.0.3
dnspython==1.16.0
Flask==2.0.2
Flask-RESTful==0.3.9
gunicorn==20.1.0
itsdangerous==2.0.1
MarkupSafe==2.0.1
peewee==3.14.4
pycodestyle==2.8.0
pymongo==3.12.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2